Conflicts:
	dpdefs/menudefs.qc
	prvm_cmds.c
	prvm_cmds.h

+++ libcurl.c
@@ -206,7 +206,7 @@ typedef struct downloadinfo_s
 	CURL *curle;
 	qboolean started;
 	int loadtype;
-	unsigned long bytes_received; // for buffer
+	size_t bytes_received; // for buffer
 	double bytes_received_curl; // for throttling
 	double bytes_sent_curl; // for throttling
 	struct downloadinfo_s *next, *prev;
@@ -430,7 +430,10 @@ static size_t CURL_fwrite(void *data, size_t size, size_t nmemb, void *vdi)
 
 	di->bytes_received += bytes;
 
-	return ret; // why not ret / nmemb?
+	return ret;
+	// Why not ret / nmemb?
+	// Because CURLOPT_WRITEFUNCTION docs say to return the number of bytes.
+	// Yes, this is incompatible to fwrite(2).
 }
 
 typedef enum
